# server/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: White-Label School Management System
  version: "3.4.0"
  description: >
    Multi-tenant school management API. Supports SuperAdmin (platform-level)
    and tenant-scoped Admin/Teacher/Student roles.
    v3.4.0: Student role added; attendance correction (PUT /attendance/{recordId});
    link-student-account (PUT /students/{studentId}/link-account);
    reactivate tenant (PUT /super-admin/tenants/{tenantId}/reactivate);
    createTenant now requires an admin block.
servers:
  - url: http://localhost:3000/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, timestamp]
          properties:
            code:
              type: string
              example: INVALID_CREDENTIALS
            message:
              type: string
              example: Invalid credentials
            details:
              type: object
              additionalProperties: true
            timestamp:
              type: string
              format: date-time
...
    SuperAdminProfile:
      type: object
      required: [id, name, email]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
    Tenant:
      type: object
      required: [id, name, slug, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, inactive]
        deactivatedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid token
              timestamp: "2026-02-26T01:00:00Z"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Insufficient permissions
              timestamp: "2026-02-26T01:00:00Z"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
...
  /super-admin/auth/login:
    post:
      summary: SuperAdmin login
      operationId: superAdminLogin
      security: []
      tags: [SuperAdmin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
            examples:
              success:
                value:
                  email: platform@admin.com
                  password: supersecret123
      responses:
        "200":
          description: SuperAdmin authenticated
          content:
            application/json:
              schema:
                type: object
                required: [token, superAdmin]
                properties:
                  token:
                    type: string
                  superAdmin:
                    $ref: "#/components/schemas/SuperAdminProfile"
              examples:
                success:
                  value:
                    token: eyJhbGciOiJIUzI1NiJ9.satoken
                    superAdmin:
                      id: SA001
                      name: Platform Admin
                      email: platform@admin.com
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─────────────────────────────────────────
  # SUPER ADMIN — TENANTS
  # ─────────────────────────────────────────
  /super-admin/tenants:
    get:
      summary: List all tenants
      operationId: listTenants
      tags: [SuperAdmin]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: object
                required: [tenants]
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tenant"
              examples:
                success:
                  value:
                    tenants:
                      - id: T001
                        name: Sunrise School
                        slug: sunrise
                        status: active
                        deactivatedAt: null
                        createdAt: "2026-01-01T00:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      summary: Create new tenant (v3.3 — seeds 8 default school_periods atomically)
      operationId: createTenant
      tags: [SuperAdmin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, name, slug, admin]
              properties:
                id:
                  type: string
                  maxLength: 50
                name:
                  type: string
                  maxLength: 255
                slug:
                  type: string
                  maxLength: 100
                  pattern: "^[a-z0-9-]+$"
                admin:
                  type: object
                  required: [name, email, password]
                  description: Initial Admin user for this tenant (v3.4 CR-06)
                  properties:
                    name:
                      type: string
                      maxLength: 255
                    email:
                      type: string
                      format: email
                    password:
                      type: string
                      minLength: 8
            examples:
              success:
                value:
                  id: T002
                  name: New School
                  slug: newschool
                  admin:
                    name: School Admin
                    email: admin@newschool.com
                    password: securepass123
      responses:
        "201":
          description: Tenant created with initial Admin user and 8 default periods
          content:
            application/json:
              schema:
                type: object
                required: [tenant, admin]
                properties:
                  tenant:
                    $ref: "#/components/schemas/Tenant"
                  admin:
                    type: object
                    required: [id, name, email, roles]
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      email:
                        type: string
                        format: email
                      roles:
                        type: array
                        items:
                          type: string
              examples:
                success:
                  value:
                    tenant:
                      id: T002
                      name: New School
                      slug: newschool
                      status: active
                      deactivatedAt: null
                      createdAt: "2026-02-26T07:00:00Z"
                    admin:
                      id: U-abc123
                      name: School Admin
                      email: admin@newschool.com
                      roles: [Admin]
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Duplicate tenant ID/slug or admin email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                admin_email_taken:
                  summary: Admin email already exists in this tenant
                  value:
                    error:
                      code: ADMIN_EMAIL_TAKEN
                      message: Admin email already exists in this tenant
                      timestamp: "2026-02-26T07:00:00Z"
                slug_conflict:
                  summary: Tenant ID or slug already taken
                  value:
                    error:
                      code: CONFLICT
                      message: Tenant ID or slug already exists
                      timestamp: "2026-02-26T07:00:00Z"
  /super-admin/tenants/{tenantId}:
    put:
      summary: Update tenant name or slug
      operationId: updateTenant
      tags: [SuperAdmin]
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                slug:
                  type: string
                  maxLength: 100
                  pattern: "^[a-z0-9-]+$"
            examples:
              success:
                value:
                  name: Updated School Name
      responses:
        "200":
          description: Tenant updated
          content:
            application/json:
              schema:
                type: object
                required: [tenant]
                properties:
                  tenant:
                    $ref: "#/components/schemas/Tenant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /super-admin/tenants/{tenantId}/deactivate:
    put:
      summary: Deactivate a tenant
      operationId: deactivateTenant
      tags: [SuperAdmin]
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant deactivated
          content:
            application/json:
              schema:
                type: object
                required: [tenant]
                properties:
                  tenant:
                    $ref: "#/components/schemas/Tenant"
              examples:
                success:
                  value:
                    tenant:
                      id: T001
                      status: inactive
                      deactivatedAt: "2026-02-26T07:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Tenant already inactive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: ALREADY_INACTIVE
                  message: Tenant is already inactive
                  timestamp: "2026-02-26T07:00:00Z"

  # v3.4 CR-07: Reactivate tenant
  /super-admin/tenants/{tenantId}/reactivate:
    put:
      summary: Reactivate an inactive tenant
      operationId: reactivateTenant
      tags: [SuperAdmin]
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Tenant reactivated
          content:
            application/json:
              schema:
                type: object
                required: [tenant]
                properties:
                  tenant:
                    $ref: "#/components/schemas/Tenant"
              examples:
                success:
                  value:
                    tenant:
                      id: T001
                      status: active
                      deactivatedAt: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Tenant already active
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: ALREADY_ACTIVE
                  message: Tenant is already active
                  timestamp: "2026-02-26T07:00:00Z"
...
  /super-admin/tenants/{tenantId}/features/{featureKey}:
    put:
      summary: Enable or disable a feature for a tenant
      operationId: toggleTenantFeature
      tags: [SuperAdmin]
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: string
        - name: featureKey
          in: path
          required: true
          schema:
            type: string
            enum: [timetable, attendance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enabled]
              properties:
                enabled:
                  type: boolean
            examples:
              enable:
                value:
                  enabled: true
              disable:
                value:
                  enabled: false
      responses:
        "200":
          description: Feature toggled
          content:
            application/json:
              schema:
                type: object
                required: [feature]
                properties:
                  feature:
                    $ref: "#/components/schemas/Feature"
              examples:
                success:
                  value:
                    feature:
                      key: attendance
                      enabled: true
                      enabledAt: "2026-02-26T07:00:00Z"
        "400":
          description: Attendance requires timetable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: FEATURE_DEPENDENCY
                  message: Attendance module requires Timetable to be enabled first
                  timestamp: "2026-02-26T07:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────────────────────
  # FEATURES (tenant Admin — read-only)
  # ─────────────────────────────────────────
  /features:
    get:
      summary: List enabled features for current tenant (Admin read-only)
      operationId: getTenantFeatures
      tags: [Features]
      responses:
        "200":
          description: Feature list
          content:
            application/json:
              schema:
                type: object
                required: [features]
                properties:
                  features:
                    type: array
                    items:
                      $ref: "#/components/schemas/Feature"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...